pack: fix packing when curtin is installed inside a snap.
When curtin is installed as part of a MAAS snap, then it would fail
to find its helpers directory.  That is because instead of helpers
living in usr/lib/curtin/helpers they will be in
 /snap/maas/<version>/usr/lib/curtin/helpers
But helpfully, SNAP will be set in the environment to point to
 /snap/maas/<version>/

The fix is to adjust pack to consider trust the environment variable
and pack up the installed files based on that prefix.

bzr-revno: 548.1.1