packaging: Update new-upstream-snapshot to work with git.
After the move to git, we need to update new-upstream-snapshot.
This version comes from cloud-init, with an extra hunk in it to
handle the transition from bzr.

Commit new-upstream-snapshot to master, where it previously
only lived on ubuntu branches.  Then, commit it also to the branches
so that it does not cause conflict.