clear-holders: detect and remove devices from kernel swap as needed.
During testing of subiquity the ephemeral environment may activate
existing swap partitions which then prevent curtin from using a device
during storage configuration. This branch introduces a change to TestBasic
to add a swap partition and a command to the dirty-disk mode to forcibly
enable the swap device to reproduce the error. The fix involves having
clear_holders detect and remove the device (disk or partition) from the
kernel swap via use of swapoff.

LP: #1743643