tests: cleanup the apt custom sources test.
This test must have gone through some iteration in development.
It was testing in a very white-box way, and not really even validating
that the config was as expected (as made evident by the previous
lack of use of EXPECTED_CONVERTED_CONTENT).

The test now asserts that the two expected changes are made.