zfs: only load zfs on xenial+ skip artful/bionic zfs env export
The zfs module is only available on xenial+, so guard against modprobing it on
precise and trusty.

Xenial is the only release which curtin is exporting an environment variable to
aid grub to find zfs devices, therefor we should only check if the environment
has the exported value on xenial, skiptest otherwise.