fix /usr/bin/curtin exit code in install failure
When curtin encounters an error during the 'install' stage, the default
configuration enables collecting of system logs.  The collect_logs function
'create_log_tarfall' included a sys.exit(0) which should only be part of the
'collect_logs_main' method; this resulted in failed installs exiting with a
return code of 0.  The fix is to move the sys.exit(0) up to the main method.

LP: #1751176