vmtest: Remove xenial and trusty specific workaround for rooturl.
xenial and trusty required specifying rc-initrd-dns on the kernel
command line in order to enable dns transition from initramfs to
real root (rooturl).  Those changes are now fixed in both releases
so we can remove this code.

The changes made the archive today.  We would expect they
are in any daily image with a serial newer than that 20180212.