pycodestyle: Fix invalid escape sequences in string literals.
A bit of information from python doc that I got by never having known:
  String literals may optionally be prefixed with a letter `r' or `R';
  such strings are called raw strings and use different rules for
  backslash escape sequences.
  ...
  Unless an `r' or `R' prefix is present, escape sequences in strings are
  interpreted according to rules similar to those used by Standard C.

So basically, any use of \ not followed by one of [\'"abfnrtv]
or \ooo (octal) \xhh (hex) or a newline is invalid.  This is most
commonly seen for us in regex.  To solve, you either:
 a.) use a raw string r'...'
 b.) correctly escape the \ that was not intended to be interpreted.

Python has deprecated these invalid string literals now
  https://bugs.python.org/issue27364
and pycodestyle is identifying them with a W605 warning.
  https://github.com/PyCQA/pycodestyle/pull/676