clear-holders: fix lvm name use when shutting down
The shutdown_lvm handler was reconstructing the name of
the lvm device incorrectly if the vg or lv name included
a hyphen.  LVM requires names with hyphens to duplicate the hyphen.
Resolve this issue by keeping the original name from the sysfs
entry and using that when issuing the dmsetup command.

LP: #1764602