zfsroot: add additional checks, do not require disk 'serial' attribute
Add some additional validation to the storage config when using 'zfsroot'
format type.  Ensure that the bootdisk is formated with GPT.  Allow a
storage config using zfsroot/zpool to work with disks without a 'serial'
attribute, but log a warning that the pool may be unstable due to changes
in device names.  Add unittests; update documentation with restrictions.

LP: #1760879
LP: #1760880