make_dname for bcache should use backing device uuid
bcache minor numbers are not stable across reboots which means
the dname symlinks pointing to a bcacheN isn't going to be stable either.
Instead, build a udev rule that links the backing device's bcache uuid
to the assigned dname.  Additionally also update the backing device
'label' sysfs attribute with the dname value.

This does rely on changes to bcache-tools to trigger the rules, currently
there is a kernel bug (LP: #1729145) which is pending, and a user-space
workaround in bcache-tools (LP: #1728742).  Either is sufficent for
the updated rules to be effective.

Some additional fixes found during this work

command/curthooks.py
  - don't append .rule to rule files, they already have .rule
examples/tests/mdadm_bcache_complex.yaml
  - removed, duplicate of mdadm_bcache.yaml