vmtests: adjust lvm_iscsi dnames to match configuration
Previous to dname/bcache changes, we failed to check the correct dname
symlinks.

In test_lvm.py we included a test_dname override to prevent checking dname
on trusty but it failed to call the parent class if we weren't skipping.

Commit 12d7163cea8, make_dname, replaced all of the per-subclass checks
for trusty and added the SkipTest to the base class implementation which
then showed the error.

This change adjusts the dnames to match what's actually generated.