pyflakes: fix unused variable references identified by pyflakes 2.0.0.
A newer version of pyflakes (2.0.0) was released over the last weekend.
It identifed some unused variables that version 1.6.0 did not identify.
The change here merely fixes those unused variables.

The fact that c-i hit this is not intended.  A future fix is needed to
properly pin the version of pyflakes so that it does not happen again.