curtainer: patch source version from --source.
After adding 'CURTIN_EXE_VERSION' and 'CURTIN_VERSION' in output
we now see nice output for runs from trunk. But when running from
a package via curtainer, we see:
  CURTIN_VMTEST_CURTIN_EXE=.. curtin-vmtest-proposed-a-29 curtin
  CURTIN_VMTEST_CURTIN_EXE_VERSION=18.1-17-gae48e86f-0ubuntu1~17.10.1
  CURTIN_VMTEST_CURTIN_VERSION=18.1

That is confusing as it appears our vmtest version is different
from our exe version.

That is because the patching of the version that we do in debian/rules
only applies to the built binaries, the source still has the un-modified
@@PACKAGED_VERSION@@ in tact.

The change here is patch the source code output by --source
by default, so that running 'curtin version' from that source will
provide the correct output.  Jenkins is the only known consumer of this
ouput so it was easier to change the default behavior than have the caller
become aware.  To disable this behavior pass '--no-patch-version'

Note that curtin has verified that it downloaded the
version of the source that it is patching in, so this is
known-to-be-correct.