tox: add a xenial environments, default envlist changes.
A previous commit used mock.assert_called, which is not
present in xenial versions of mock so build would fail on xenial.
Due to bugs that are present in the trusty versions of mock it would
pass there.

So we add a tox environment 'xenial-py3' and add that to the
default set of environments that are run with a 'tox' invocation.

Other changes to the envlist here:
a.) make it multiple lines for easier future diffing.
b.) run py3-flake8 first, it runs quickly and gives good feedback.
c.) drop trusty-check which was running trusty pyflakes.  We do not
    run pyflakes in package build any more.
d.) replace trusty-py3 with xenial-py3. Just to save time in default 'tox'
    run we only run one "old" python3 unit tests.