Fix extraction of local filesystem image.
Curtin would fail to extract an image if the image was a
local filesystem path or a file:// url.

Also here is a fix for ensure_dir which would be tripped if you
call write_file("foo.txt").  In that path,  ensure_dir would be called
with the result of os.path.dirname("foo.txt") which is "" and
os.makedirs("") would fail.

LP: #1775630