vmtests: Increase root fs to give upgrades to -proposed more space.
Our new jenkins job curtin-vmtest-devel-amd64-proposed runs with
CURTIN_VMTEST_ADD_REPOS=proposed . That causes an
'apt-get update && apt-get dist-upgrade' to occur in the installed
system.

The change here simply increases by 500M the size of the root
disk tests that use examples/tests/mirrorboot-msdos-partition.yaml.
We can simply do this because the md0-part1 was on md0 which was a
5G disk, but we were creating 2x2G partitions.  So now we still have
some wiggle room there.

This particular build failed because of 2 reasons
a.) the HWE variety meant additional kernels installed
b.) the trusty images are well out of date at thsi point (2018-05-02)
so lots of apt downloads and upgrades.