vmtests: Let a raised SkipTest go through skip_by_date.
There was a general logic error in skip_by_date.  If a setUpClass
raised a SkipTest and the fix-by date had not been raised then
we would swallow that SkipTest.  Th result was that the test methods
would fail as the install had not been done.

We saw this specifically at the end of Artful's support.  The
artful setUpClass would then raise SkipTest due to
'is_unsupported_ubuntu'.  skip_by_date was swallowing that.

The change here is to just let a SkipTest bubble through the
skip_by_date wrapper.