clear-holders: reread ptable after wiping disks with partitions
When wiping a disk with partitions curtin races the kernel in removing
sysfs entries for the partitions.  This shows up as a OSError when
asserting the disk is clear as partition sysfs paths are unexpectedly gone
during the building of the holder tree.

To prevent the race, wipe_superblock will check if the device has
partitions and if so after wiping the the superblock, rescan the
device partition table and then check that the device no longer has
partitions, possibly sleeping and rescanning.