vmtest: enable persistent journal and collect at boot time
Target OSes which use systemd-journald may or maynot enable a
persistent journal on the root filesystem.  If the target release
supports journald, issue the commands to create a persistent
journal.  During the boot phase, issue commands to flush the
journal and collect (and compress) the journal as the last
script before tar'ing the collect data.

Rename vmtest_pollinate.yaml to vmtest_defaults.yaml.  This
file is included in all vmtest runs.