vmtest: ensure we collect /var/log/journal only once
Appending the journal script to the collect_scripts parameter
modified a share class variable which resulted in more than
one call to the script to collect the journal.  The script itself
was not defensive about repeated calls and would copy the
journal dir more than once which resulted in filling the capacity
of the output disk.

The changes here make the script defensive to only run once,
and fail if additional attempts to run are made.