vmtests: prevent tests from modifying cls.collect_scripts
vmtests now has a set of collect scripts which are shared in all
subclassed tests and should not be modifed by the subclass.  Fix
test_lvm_raid.py which modified cls.collect_scripts to use
extra_collect_scripts and update the base class to raise a RuntimeError
if a subclass presents a modified cls.collect_scripts.