Capture stdout when using lxc file push
Newer lxd (3.4+) 'file' command now produces progress info in some
cases written to stdout.  This breaks curtin-from-container which
is invoked with the 'pack' curtin command which also writes to stdout.
The result was a corrupt curtin package output file which left all
vmtests sitting idle in the ephemeral boot phase and timing out.

To resolve, we capture stdout from lxc file push and report
the contents if file push fails.