vmtests: dont use multiple subclasses in uefi 4k tests
The uefi 4k sector size test used an intermediary subclass
to set the sector_size of 4k, but the resulting subclass
ended up creating an instance of the parent class which set
the 'collect_scripts' attribute and then tripped the runtime
check on vmtests having an empty 'collect_scripts' attribute.

The fix was to create the 4k tests directly from the base class.