vmtests: fix common collect scripts to not exit failure.
If collect scripts exit non-zero, then cloud-init in the booted system
will complain that they have done so.  This adds noise to logs.

At this point, the only command in a long list that affects the exit code
is the last one.  At some point in the future we may have more systematic
collection.  At that point it might make sense to exit if the collection
failed.

For now, just exit 0.