vmtests: multipath mount /home with nofail and validate in unittest
In some scenarios the mounting of /home partition under multipath fails
and leaves the boot stage hung in emergency shell.  We have no ability
to debug this in automated runs on infrastructure.  Instead we will now
pass the 'nofail' mount option which instructs systemd to not fall to
the emergency shell allowing the boot to complete and vmtest to collect
information.  We now validate the status of the home mount in the
unittest in the class.  When this fails, we will have collected the
journal and the unit information which should help narrow down why
we see the failure.