vmtests: check install log for Out of memory kernel messages and fail
During proposed testing, we observed in the Cosmic test which installed
the lxd snap, core snap and a linux kernel, 1G of ram was insufficient
and would trigger linux OOM and sometimes fail the install.  Add a
test for the Out of memory string in our install log check and fail
vmtest installs if we find an OOM occurred during installation.