Fix bug in is_swap_device if a device was smaller than page_size.
If is_swap_device was called on a device (or file) that was smaller than
a page size of the current kernel, then it would raise an OSError.
The change here just makes it check the size of the file before
attempting to read.

A test is added, and the other tests updated to rely less on mock.

LP: #1803672