curthooks: implement chzdev export, import for active or online devices
For s390x systems, curtin is responsible for enabling persistent device
configuration in the target.  This entirely handled by use of the s390
tool, chzdev, using the --export and --import feature to emit the
chzdev configuration format, modifying the output to be used to import
and configure the active or online devices persistently in the target.