clear-holders: handle FileNotFound when probing for bcache device slaves
In some cases, when detecting if a bcache has any slave devices which
need removal the path to the sysfs directory may not be found; this is
non-fatal; it indicates that the existing device has been removed by the
kernel.  Allow clear-holders to tolerate this failure.

Updated curtin.block.partition_kname to indicate that bcache partitions
have 'p' in the value.  This is needed for the vmtest to recreate bcache
partitions.

LP: #1811117