bcache: ensure partitions on bcache devices are detected as partition
In some cases, when curtin is discovering existing storage devices, a
bcache device with partitions created on top of it, may be recognized
as a bcache device, rather than a partition.  If curtin treats a
bcache partition as a bcache device, it will attempt to write to
non-existent files on /sys which results in a failure after timing
out.  Resolve this issue by detecting if the target device is a
partition inside the bcache identity function.