Support for multi-layers images fsimage-layered:// URI
Curtin can now download and mount a layered filesystem image for use
as a source image.  Local file or url are supported. Filesystem can be
any filesystem type mountable by the running kernel.

A "fsimage-layered" install source is a string representing one or
more mountable images from a single local or remote directory.  The
string is dot-separated where each value between the dots represents a
particular image and the location of the name within the string
encodes the order in which it is to be mounted.  The resulting list of
images are downloaded (if needed) then mounted and overlayed into a
single directory which is used as the source for installation.

See documentation for further details.