clear-holders: wipe raid members at data offsets when not assembled
RAID5 arrays are assembled in degraded mode for speed[1].  If an
array is INACTIVE, then curtin cannot wipe the array through the
md device.  Instead, examine the array members and find the mdadm
superblock and data offsets and wipe members at those locations.
This wipes the array contents without requiring it to be runnable.

LP: #1815018