vmtests: add support for CURTIN_VMTEST_APT_PROXY
This has the same behaviour as apt_proxy, but allows developers to
override it globally without affecting other applications. apt_proxy
will continue to work, but CURTIN_VMTEST_APT_PROXY will be preferred if
it is present in the environment.

(This change is applied to both the Python vmtests code, and to the
tools/launch shell script.)