block: Add opportunistic zkey encryption if supported
On s390x, systems with a crypto accelerator may be present
and enabled for sure.  When handling a type: dm_crypt block
configuration, curtin will test if zkey is available and if
so, use the zkey command to generate keys and encrypt the
block device using zkey-based secrets.

In the case that zkey is not available, curtin will fallback
to using normal cryptsetup.