util.ChrootableTarget: skip rename of resolv.conf if not present in target
A target OS may not include an /etc/resolv.conf. ChrootableTarget attempts
to move the in-chroot resolv.conf out of the way and copy in the host
resolv.conf. If the target image does not have /etc/resolv.conf then we
fail when we call os.rename. Avoid this error by only invoking the
rename if the target image has a resolv.conf.

LP: #1834382