net: ensure eni helper tools install if given netplan config
Curtin did not check if the target system had the require packages
assuming that v2 network config would only be sent to bionic and
newer.  We can handle this by emitting dependencies on bridge, vlan
and bond configurations.

Add a vmtest to verify this.  In doing so, the vmtest requires that
we parse the provided network-config file, so add support for netplan
config parsing to curtin/net/network_state.py.  Also note that Xenial
bridge-utils does not handle per-path settings like path-cost or
path-priority; so we disable verifying those values in the test.

LP: #1834751