vmtests: use file locking on the images
This change implements a file locking mechanism in the
jenkins-runner and vmtest-sync-images scripts.

The vmtest-sync-images script requires two exclusive locks
to start updating the images: a lock on IMAGE_DIR and a
lock on a lockfile. The script will wait to get both locks
locks before actually updating the images.

The jenkins-runner script does the following:

 (1) set a shared lock on IMAGE_DIR
 (2) set a shared lock on the lockfile
 (3) release the lock on IMAGE_DIR
 (4) run the vmtests
 (5) exit (releasing the lock on the lockfile)

The jenkins-runner does not hold a lock on IMAGE_DIR while
running, allowing vmtest-sync-images to set a lock there,
prevending further jenkins-runners from running the vmtests
while the images are being updated. In other words
vmtest-sync-images can "book" the images for the next
available slot.

Steps (1) and (3) can be skipped by passing the
--skip-images-dirlock option to jenkins-runner. In this way
jenkins-runner has precedence over vmtest-sync-images. This
is useful when it should run ASAP, for example in CI runs.