jenkins-runner: disable the lockfile timeout by default
When CURTIN_VMTEST_IMAGE_LOCK_TIMEOUT is set to a negative integer
disable the lockfile timeout. By default set it to -1 (no timeout).

Rationale: while it is useful to have a timeout, there is no
sensible default. In a run like (parameters omitted):

  $ jenkins-runner &
  $ vmtest-sync-images &
  $ jenkins-runner

the second jenkins-runner will have to wait for the first one to
finish, as it's locked by vmtest-sync-images. A full vmtest run
can take several hours. I easily spotted successful runs that
took more than 16 hours on our Jenkins log. Setting the default
timeout to something like 20 hours should be safe, but I still
don't think it's a good default, as it is too high for all the
cases where only a subset of the tests is being run. The timeout
is too dependent on the hardware, on how busy the server is,
and on the set of tests which is run. No size fits all.

Other minor changes:
 - Force CURTIN_VMTEST_IMAGE_LOCK_TIMEOUT to be an integer
 - Remove unnecessary quotes around fixed strings