Handle partial raid on partitions
Consider disks and partitions when checking for holders that
need removal before they can be used.  Additionally when shutting
down a raid, if it is inactive or broken, writes to the /dev/mdX
device will fail which is expected, don't exit  instead, log a
non-fatal error and continue with shutting down the raid.

With these fixes TestReuseRAIDMemberPartition vmtests now pass.

LP: #1835091