clear-holders: increase the level for devices with holders by one
In the case where clear-holders looks at a dependent device, if the
parent is already in the registry, we want to take the max level
and then increment this by one to ensure the dependent device is
shutdown first.  This resolves the case where we have an LVM over
top a bcache device and we need to remove the LVM device before
any of the bcache devices can be removed as they  may share a
cacheset in which case all bcache devices must be stopped before
the cacheset can be removed.

LP: #1844543