block-schema: update raid schema for preserve and metadata
Curtin can now accept type: raid storage config which sets the
'preserve' boolean to 'true' which informs curtin that it does
not need to create a new raid device.

We also accept a 'metadata' version hint in the schema as curtin
passes this value along as info to users of block-discover.