tools/xkvm: add -nographic to speed up devopt query
On systems which include qemu graphical components the time
needed to invoke qemu for querying if an string is a device
option can take many seconds, especially over remote connections.
Adding the '-nographic' parameter prevents qemu from loading
and launching additional resources that are not needed to determine
the devopt query answer.