vmtest: Adjust TestScsiBasic to use dnames to find correct disk
The TestScsiBasic test has some assumptions on which SCSI device
will have expected content. On PPC64, the PCI device enumeration
is not as stable as it is on x86 resulting in test failures. This
patch resolves this by leveraging the dname values and looking up
the correct kernel name for a specific target dname.