block-discover: handle partial probe data
The probe_data provided to curtin may only include a subset of
the types of data that curtin knows how to parse.  Currently if
one of the types is not present in the probe_data, curtin raises
Exceptions which prevent callers from getting any storage config
from curtin.  This branch allows curtin to handle missing data
and will log the errors but not raise Exceptions, leaving the
caller to handle and decide what to do with the storage config
produced.

LP: #1852351