curthooks: only run update-initramfs in target once
During an installation of Ubuntu; various packages (linux-image,
zfs-initramfs, mdadm, multipath-tools, etc) require an update
to the initramfs contents to ensure configuration details are
available.  In the worst-case we call update-initramfs three
or more times.  This branch disables update-initramfs in the
target system by moving it out of the way; then it replaces
the update-initramfs binary with equivalent of /bin/true as
some package's post-install hook call update-initramfs directly
and fail without it being present.  At the end of curthooks
curtin re-enables the original update-initramfs and ensures
that all kernels installed will have an initrd.img created
or updated as needed.

Additional changes:
- block_meta
  - enable strict=True loading of command environment state
    to ensure curtin has required state present.
  - Replace open-coded file writing with util.write_files
    for mdadm and crypttab.
- curthooks
  - Use dpkg-divert to check if update-initramfs has been
    diverted.
  - Move update-initramfs actions (redhat & debian) under
    the same event reporter.
- vmtests
  - Don't unset OUTPUT_FSTAB, instead provide an alternative
    location.
  - Move class-based extra_kernel_args to the end; some
    class configs specify '---' which affects the
    carry-over kernel parameters used when booting the
    installed image.

LP: #1842264